python/mcp-mini/pyproject.toml
[project]
name = "mcp-mini"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic==2.11.7",
    "python-dotenv>=1.2.1",
    "requests==2.32.5",
]
